requirements.txt
fastapi
uvicorn[standard]
streamlit
pandas
statsmodels
scipy
plotly
sqlalchemy
requests
websocket-client